requirements.txt
Django==2.2.12
djangorestframework==3.10.2
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0
django-notifications-hq==1.6.0